% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.SI.R
\name{AIC.surveyIdx}
\alias{AIC.surveyIdx}
\title{Akaike Information Criterion (or BIC) for survey index models}
\usage{
\method{AIC}{surveyIdx}(x, BIC = FALSE)
}
\arguments{
\item{x}{survey index as return from getSurveyIdx}

\item{BIC}{if TRUE compute BIC instead of AIC}
}
\value{
numeric value
}
\description{
Akaike Information Criterion (or BIC) for survey index models
}
