% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.SI.R
\name{anova.SI}
\alias{anova.SI}
\title{Likelihood ratio test for comparing two survey indices.}
\usage{
\method{anova}{SI}(m1, m2)
}
\arguments{
\item{m1}{}

\item{m2}{}
}
\value{
A p-value.
}
\description{
Likelihood ratio test for comparing two survey indices.
}

