% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concTransform.R
\name{concTransform}
\alias{concTransform}
\title{Helper function for plotting survey indices.}
\usage{
concTransform(x)
}
\arguments{
\item{x}{a vector of log-responses}
}
\value{
vector of transformed responses
}
\description{
Concentration transform
}

