% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIdxPlots.R
\name{depthDist}
\alias{depthDist}
\title{Calculate and plot the distribution as a function of depth}
\usage{
depthDist(
  m,
  grid,
  by = 5,
  type = c("dist", "mean"),
  colfun = colorRampPalette(c("red", "lightgrey", "blue")),
  lwd = 1.5,
  rangeFrac = 0.5,
  age = 1,
  ...
)
}
\arguments{
\item{m}{object of class 'surveyIdx'}

\item{grid}{grid object used when fitting 'm'}

\item{by}{depth bin size}

\item{type}{character vector indicating what to plot. 'dist' or 'mean' or both.}

\item{colfun}{color function to use for plotting.}

\item{lwd}{plotting line width}

\item{rangeFrac}{Fraction ]0;1[ indicating the percentile range in the mean plot. Default is 0.5, i.e. interquartile range.}

\item{age}{which age group to plot}

\item{...}{extra arguments to 'plot}
}
\value{
a matrix with depth distribution in each year.
}
\description{
Calculate and plot the distribution as a function of depth
}
