% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSI.R
\name{exportSI}
\alias{exportSI}
\title{Write survey index to file in standard XSA/SAM format}
\usage{
exportSI(x, ages, years, toy, file, nam = "")
}
\arguments{
\item{x}{matrix with survey indices}

\item{ages}{vector of ages}

\item{years}{vector of years}

\item{toy}{fraction of year the survey is conducted (between 0 and 1)}

\item{file}{filename to write to}

\item{nam}{file description header}
}
\value{
nothing
}
\description{
Write survey index to file in standard XSA/SAM format
}
