% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCons.R
\name{externalCons}
\alias{externalCons}
\title{Calculate external consistencies between two survey indices.}
\usage{
externalCons(tt, tt2, do.plot = FALSE)
}
\arguments{
\item{tt}{A matrix with survey indices (rows=years, cols=ages)}

\item{tt2}{A matrix with survey indices (rows=years, cols=ages)}

\item{do.plot}{plot it?}
}
\value{
A vector of correlations (consistencies)
}
\description{
Calculate external consistencies between two survey indices.
}
\details{
Proper alignment of years and ages must be ensured by the user.
}
