% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAgeGroup.R
\name{fixAgeGroup}
\alias{fixAgeGroup}
\title{Helper function to "borrow" missing age groups from other years}
\usage{
fixAgeGroup(x, age = 0, n = 3, fun = "mean")
}
\arguments{
\item{x}{DATRASraw object}

\item{age}{age to impute}

\item{n}{at least this many individuals in each year}

\item{fun}{A function such as 'mean','median','min', or 'max'.}
}
\value{
a DATRASraw object
}
\description{
Helper function to "borrow" missing age groups from other years
}
\details{
In years where there are less than 'n' individuals of age 'age',
add fake individuals of that age such that there are 'n'.
The length of the individuals are set to the mean (or whatever 'fun' specifies)
of all other individuals of the same age.
For the minimum and maximum age groups fun it is reasonable to replace 'mean' with 'min' and 'max' respectively.
Note, that you might need to call 'addSpectrum' on the object again.
}
