% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBathyGrid.R
\name{getBathyGrid}
\alias{getBathyGrid}
\title{Get bathymetric prediction grid corresponding to the area for a DATRASraw object using the 'marmap' package}
\usage{
getBathyGrid(
  d,
  minDepth = 10,
  maxDepth = Inf,
  resolution = 2,
  maxDist = Inf,
  keep = TRUE,
  shapefile = NULL,
  select = NULL
)
}
\arguments{
\item{d}{DATRASraw object}

\item{minDepth}{Minimum depth to include}

\item{maxDepth}{Maximum depth to include}

\item{resolution}{grid resolution (see marmap::getNOAA.bathy)}

\item{maxDist}{Do not include grid points farther than maxDist from nearest observation.}

\item{keep}{Save grid on disk for fast loading next time?}

\item{shapefile}{extra shapefile information to add (optional)}

\item{select}{columns to extract from shapefile}
}
\value{
data.frame with depths and geographical coordinates
}
\description{
Get bathymetric prediction grid corresponding to the area for a DATRASraw object using the 'marmap' package
}
