% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrid.R
\name{getGrid}
\alias{getGrid}
\title{Create a grid of haul positions from a DATRASraw object.}
\usage{
getGrid(dd, nLon = 20)
}
\arguments{
\item{dd}{DATRASraw object}

\item{nLon}{number of grid cells in the longitude direction.}
}
\value{
A surveyIndexGrid (a list of coordinates and haul.ids)
}
\description{
Create a grid of haul positions from a DATRASraw object.
}
