% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSurveyIdxStratMean.R
\name{getSurveyIdxStratMean}
\alias{getSurveyIdxStratMean}
\title{Survey index using the stratified mean method using ICES statistical rectangles as strata.}
\usage{
getSurveyIdxStratMean(x, ageCols, doLog = FALSE)
}
\arguments{
\item{x}{DATRASraw object. Must contain a matrix: x[[2]]$Nage.
##'}

\item{ageCols}{which columns of the Nage matrix should be included?}

\item{doLog}{log-transform?}
}
\value{
a matrix with survey indices
}
\description{
Survey index using the stratified mean method using ICES statistical rectangles as strata.
}

