% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internalCons.R
\name{internalCons}
\alias{internalCons}
\title{Calculate internal consistency of a survey index.}
\usage{
internalCons(tt, do.plot = FALSE)
}
\arguments{
\item{tt}{A matrix with survey indices (rows=years, cols=ages)}

\item{do.plot}{Plot it?}
}
\value{
a vector of consistencies
}
\description{
Calculate internal consistency of a survey index.
}

