% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroLO.R
\name{leaveout.surveyIdx}
\alias{leaveout.surveyIdx}
\title{Make leave-one-out analysis}
\usage{
leaveout.surveyIdx(model, d, grid, fac)
}
\arguments{
\item{model}{object of class "surveyIdx" as created by "getSurveyIdx"}

\item{d}{DATRASraw dataset}

\item{grid}{surveyIndexGrid object (see getGrid) defining the grid.}

\item{fac}{a factor in d to leave out one at a time, e.g. d$Survey}
}
\value{
SIlist (list of surveyIndex objects)
}
\description{
Make leave-one-out analysis
}
