% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIdxPlots.R
\name{mapLegend}
\alias{mapLegend}
\title{Plot values for each color level in a map as created by 'surveyIdxPlots'.}
\usage{
mapLegend(
  x,
  year,
  colors = rev(heat.colors(6)),
  age = 1,
  legend.signif = 3,
  cutp = NULL,
  lwd = 10,
  las = 2
)
}
\arguments{
\item{x}{object of class 'surveyIdx'}

\item{year}{vector of years (should be the same as used for 'surveyIdxPlots')}

\item{colors}{vector of colors to use}

\item{age}{age group (integer >=1)}

\item{legend.signif}{number of significant digits shown}

\item{cutp}{optional custom vector of cut points for color scales.}

\item{lwd}{line width for bars}

\item{las}{orientation of x-axis lables}
}
\description{
Plot values for each color level in a map as created by 'surveyIdxPlots'.
}
