% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroLO.R
\name{mohn.surveyIdx}
\alias{mohn.surveyIdx}
\title{Mohn's rho for retrospective analysis}
\usage{
mohn.surveyIdx(x, base, rescale = FALSE)
}
\arguments{
\item{x}{SIlist as returned by the 'retro.surveyIdx' function}

\item{base}{Object of class 'surveyIdx' (full run with all years)}

\item{rescale}{should the indices for each age group be rescaled to have mean 1 over shortest timespan before calculating mohns rho? Only appropriate for purely relative indices.}
}
\value{
vector with mohn's rho for each age group
}
\description{
Mohn's rho for retrospective analysis
}
