% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrid.R
\name{plot.surveyIndexGrid}
\alias{plot.surveyIndexGrid}
\title{Plot a surveyIndexGrid}
\usage{
\method{plot}{surveyIndexGrid}(grid, pch = 1, gridCol = "lightgrey")
}
\arguments{
\item{grid}{a surveyIndexGrid (as created by the "getGrid" function)}
}
\value{
nothing
}
\description{
Plot a surveyIndexGrid
}
