% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{qres.tweedie}
\alias{qres.tweedie}
\title{Randomized quantile residuals for Tweedie models}
\usage{
qres.tweedie(gam.obj)
}
\arguments{
\item{gam.obj}{A gam object (mgcv package)}
}
\value{
A vector of residuals, which should be iid standard normal distributed
}
\description{
Randomized quantile residuals for Tweedie models
}
