% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyIdx.R
\name{redoSurveyIndex}
\alias{redoSurveyIndex}
\title{Re-compute standardized survey indices for an alternative grid from a previous fitted "surveyIdx" model.}
\usage{
redoSurveyIndex(
  x,
  model,
  predD = NULL,
  myids,
  nBoot = 1000,
  predfix = list(),
  mc.cores = 1
)
}
\arguments{
\item{x}{DATRASraw dataset}

\item{model}{object of class "surveyIdx" as created by "getSurveyIdx"}

\item{predD}{optional DATRASraw object, defaults to NULL. If not null this is used as grid.}

\item{myids}{haul.ids for grid}

\item{nBoot}{number of bootstrap samples used for calculating index confidence intervals}

\item{predfix}{optional named list of extra variables (besides Gear, HaulDur, Ship, and TimeShotHour),  that should be fixed during prediction step (standardized)}

\item{mc.cores}{mc.cores number of cores for parallel processing}
}
\value{
An object of class "surveyIdx"
}
\description{
Re-compute standardized survey indices for an alternative grid from a previous fitted "surveyIdx" model.
}
