% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeZeroClusters.R
\name{removeZeroClusters}
\alias{removeZeroClusters}
\title{Remove factor levels with only zero observations, which can cause estimation problems}
\usage{
removeZeroClusters(
  d,
  cutat = 0,
  response = NULL,
  factors = c("Gear", "Ship", "Year", "StatRec"),
  verbose = TRUE
)
}
\arguments{
\item{d}{DATRASraw object}

\item{cutat}{numeric in the interval [0,1[. If greater than 0, then factor levels where mu/max(mu) < cutat are removed. Note, that this will remove some positive observations.}

\item{response}{Character naming the response column. If NULL, then total numbers are used as response.}

\item{factors}{vector naming the factors to consider.}

\item{verbose}{print out extra info? Default: TRUE}
}
\value{
DATRASraw object
}
\description{
Remove factor levels with only/many zero observations, which can cause estimation problems
}
