% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals.surveyIdx}
\alias{residuals.surveyIdx}
\title{Randomized quantile residuals for class 'surveyIndex'}
\usage{
\method{residuals}{surveyIdx}(x, a = 1)
}
\arguments{
\item{x}{An object of type 'surveyIndex' as created by 'getSurveyIdx'}

\item{a}{age group}
}
\value{
A vector of residuals, which should be iid standard normal distributed
}
\description{
Randomized quantile residuals for class 'surveyIndex'
}
