% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{surveyIdx.simulate}
\alias{surveyIdx.simulate}
\title{Simulate data from a surveyIndex model (experimental and subject to change)}
\usage{
surveyIdx.simulate(model, d, sampleFit = FALSE, condSim = NULL)
}
\arguments{
\item{model}{object of class 'surveyIdx'}

\item{d}{A dataset (DATRASraw object)}

\item{sampleFit}{Use a random sample from the gaussian approximation to distribution of the estimated parameter vector. Default: FALSE.}

\item{condSim}{optional results of previous call to this function. Use this if you want to generate many datasets (much faster, since mean predictions are re-used).}
}
\value{
list with  1) simulated observations with noise 2) mean (no noise) 3) zero probability.
}
\description{
Simulate data from a surveyIdx model (experimental and subject to change)
}
