% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIdxPlots.R
\name{surveyIdxPlots}
\alias{surveyIdxPlots}
\title{Visualize results from a survey index model fitted with getSurveyIdx().}
\usage{
surveyIdxPlots(
  x,
  dat,
  alt.idx = NULL,
  myids,
  cols = 1:length(x$pModels),
  select = c("index", "map", "residuals", "fitVsRes"),
  par = list(mfrow = c(3, 3)),
  colors = rev(heat.colors(6)),
  map.cex = 1,
  plotByAge = TRUE,
  legend = TRUE,
  predD = NULL,
  year = NULL,
  main = NULL,
  legend.signif = 3,
  legend.pos = "topright",
  restoreOldPar = FALSE,
  scaleMap = TRUE,
  posProb = FALSE,
  ...
)
}
\arguments{
\item{x}{Survey index as produced by getSurveyIndex()}

\item{dat}{DATRASraw object}

\item{alt.idx}{optional matrix with alternative index}

\item{myids}{vector of haul ids that constitute the grid}

\item{cols}{which age columns to consider?}

\item{select}{character vector of chosen plots. Either one of "index","map","absolutemap","residuals","fitVsRes",""resVsYear","resVsShip","spatialResiduals", or a number. Numbers refer to smooths in the order they appear in the formula.}

\item{par}{'par' settings for plotting (a named list).}

\item{colors}{colors for spatial effect.}

\item{map.cex}{size of grid points on maps}

\item{plotByAge}{boolean (default=TRUE). If true, par(par) is called for each age group.}

\item{legend}{boolean (default=TRUE). add legends to plot?}

\item{predD}{DATRASraw object with grid (optional). Overrides 'myids' if supplied.}

\item{year}{numeric scalar or vector (default=NULL). If 'select' equals 'map' a specific year can be chosen (only meaningful for time-varying spatial effects). If select equals 'absolutemap' year must be a vector.}

\item{main}{optional main title (overrides default title)}

\item{legend.signif}{Number of significant digits in map legends}

\item{legend.pos}{Position of legend (e.g. "bottomleft") see ?legend}

\item{restoreOldPar}{restore old par() on exit? Default=FALSE}

\item{scaleMap}{scale legend to mean 1 for "absolutemap"? Default: TRUE}

\item{posProb}{show probability of non-zero value for "absolutemap" instead of mean}

\item{...}{Additional parameters for plot()}
}
\value{
nothing
}
\description{
Visualize results from a survey index model fitted with getSurveyIdx().
}
