% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroLO.R
\name{plot.SIlist}
\alias{plot.SIlist}
\title{Plot survey index list (e.g. retrospective analysis)}
\usage{
\method{plot}{SIlist}(
  x,
  base = 1,
  rescale = FALSE,
  lwd = 1.5,
  main = NULL,
  allCI = FALSE,
  includeCI = TRUE,
  ylim = NULL,
  density = FALSE
)
}
\arguments{
\item{x}{(named) list of "surveyIdx" objects for example from "retro.surveyIdx" or "leaveout.surveyIdx"}

\item{base}{Either index of x that should considered the "base run" (integer), OR object of class "surveyIdx". Confidence bounds will be shown for this model only.}

\item{rescale}{Should indices be rescaled to have mean 1 (over the set of intersecting years)? Default: FALSE}

\item{lwd}{line width argument to plot}

\item{main}{if not NULL override main plotting default title of "Age group a"}

\item{allCI}{show 95\% confidence lines for all indices? Default FALSE.}

\item{includeCI}{Show confidence intervals? Default TRUE.}

\item{ylim}{Y axis range. If NULL (default) then determine automatically.}

\item{density}{if TRUE, divide each index by the number of grid points to get densities rather than totals.}
}
\value{
nothing
}
\description{
Plot survey index list (e.g. retrospective analysis)
}
