% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroLO.R
\name{retro.surveyIdx}
\alias{retro.surveyIdx}
\title{Make retrospective analysis for a "surveyIdx" model.}
\usage{
retro.surveyIdx(model, d, grid, predD = NULL, npeels = 5, ...)
}
\arguments{
\item{model}{object of class "surveyIdx" as created by "getSurveyIdx"}

\item{d}{DATRASraw dataset}

\item{grid}{surveyIndexGrid object (see getGrid) defining the grid.}

\item{npeels}{number of years to successively peel of the data set}

\item{...}{Optional extra arguments to "gam"}
}
\value{
SIlist (list of surveyIndex objects)
}
\description{
Make retrospective analysis for a "surveyIdx" model.
}
