% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIdxPlots.R
\name{surveyIdxPlots}
\alias{surveyIdxPlots}
\title{Visualize results from a survey index model fitted with getSurveyIdx().}
\usage{
surveyIdxPlots(x, dat, alt.idx = NULL, myids, cols = 1:length(x$pModels),
  select = c("index", "map", "residuals", "fitVsRes"), par = list(mfrow =
  c(3, 3)), colors = rev(gray.colors(5)), map.cex = 1, plotByAge = TRUE,
  legend = TRUE, predD = NULL, year = NULL, ...)
}
\arguments{
\item{x}{Survey index as produced by getSurveyIndex()}

\item{dat}{DATRASraw object}

\item{alt.idx}{optional matrix with alternative index}

\item{myids}{vector of haul ids that constitute the grid}

\item{cols}{which age columns to consider?}

\item{select}{character vector of chosen plots. Either one of "index","map","residuals", or "fitVsRes" or a number. Numbers refer to smooths in the order they appear in the formula.}

\item{par}{'par' settings for plotting (a named list).}

\item{colors}{colors for spatial effect.}

\item{map.cex}{size of grid points on maps}

\item{plotByAge}{boolean (default=TRUE). If true, par(par) is called for each age group.}

\item{legend}{boolean (default=TRUE). add legends to plot?}

\item{predD}{DATRASraw object with grid (optional). Overrides 'myids' if supplied.}

\item{year}{numeric (default=NULL). If 'select' equals 'map' a specific year can be chosen (only meaningful for time-varying spatial effects).}

\item{...}{Additional parameters for plot()}
}
\value{
nothing
}
\description{
Visualize results from a survey index model fitted with getSurveyIdx().
}
